Started writing code to compute a DAG of worlds (rather, world islands)
The code is intended to support clustering so that the huge world for
something like Prelude can be clustered into a single node.
Unfortunately, this is not currently working properly. Partly that's due
to the fact that -- ARGH -- the imports we compute from .hi files are
not actually the syntactically written imports; they also include some
modules in the transitive closure of imports. That makes it harder to
cleanly say "cluster any imports of Prelude" because each "import
Prelude" in a module will be saved as "import base:Prelude; import
base:GHC.Classes; import Base:..." etc.

After experimenting with this sloppy approach, I think the best thing to
do is cluster worlds by *package*. So each included package will have
its own world cluster in the DAG (if desired), allowing us to omit
"base", "ghc-prim", etc. from the DAG.