Allow inconsistent worlds and record consistency in World
Since we want to calculate consistency stats for packages, we want to
allow inconsistent worlds but we also want to know when a world is
consistent. So now instead of rejecting a merge of incompatible worlds,
we just record whether the merge was well-defined (and whether the
merged worlds were all consistent).