Tracking which modules produce inconsistency when reading worlds
Consistency is tracked as a list of module names (Moduleish) that merged
together parent worlds which were overlapping. (We assume that freshly
defined instances never overlap with parent worlds because it had to
pass GHC typechecking.) Now you can track the accumulated list of
modules that produced inconsistency for the given (package or module)
world.