Added another test to check how intentional overlap is handled
Right now my code doesn't detect an inconsistent world in
Test3.OverlapShow(.Client), when it obviously should. The problem is
that I'm not currently checking whether instances added in a new island
are overlapping with instances in the island's parents. This is a pretty
obvious oversight on my part. (I was relying on the well-typedness of
the module to be good enough to assert that the instances don't overlap,
but overlap is indeed allowed, presumably due to some combination of
instance flags.)

The next thing to do is to change the newWorld code to check whether the
newly added instances overlap with the parent world.