Executables now take package lists from STDIN; adding orphan results
Now the calc-worlds and orphans-stats executables take in the list of
packages to analyze from STDIN. The input format assumes each line is a
package name (with no extra whitespace), an empty line, or a comment
that begins with '#'.

Added the results of orphans-stats on platform packages, popular
packages, and all combined.