Tweaks and NEW PROBLEMS
PROBLEM 1: Discovered that old analysis didn't actually cover all
packages because I didn't have all selected packages installed???
Dunno how that happened. Now all packages are installed, but...

PROBLEM 2: There's some kernel panic related to a boot file:

```
calc-worlds: panic! (the 'impossible' happened)
  (GHC version 7.8.4 for x86_64-apple-darwin):
  lookup failed for module crypto-api-0.13.2:Crypto.Modes[boot]
```

PROBLEM 3: I realized that I'm not "tying the recurive knot" with
worlds. That means that a module B that imports a boot file A' but not
the implementing module A is only seeing the world of A'. There
needs to be a later step of tying the knot so that that B sees the
entire world of A.

Oh well, back to writing the thesis for now!